+++ net/sched/cls_flower.c
@@ -2355,7 +2355,7 @@ static int fl_ht_insert_unique(struct cls_fl_filter *fnew,
 }
 
 static int fl_change(struct net *net, struct sk_buff *in_skb,
-		     struct tcf_proto *tp, unsigned long base,
+		     struct tcf_proto *tp, uintptr_t base,
 		     u32 handle, struct nlattr **tca,
 		     void **arg, u32 flags,
 		     struct netlink_ext_ack *extack)
@@ -3810,8 +3810,8 @@ nla_put_failure:
 	return -EMSGSIZE;
 }
 
-static void fl_bind_class(void *fh, u32 classid, unsigned long cl, void *q,
-			  unsigned long base)
+static void fl_bind_class(void *fh, u32 classid, uintptr_t cl, void *q,
+			  uintptr_t base)
 {
 	struct cls_fl_filter *f = fh;
