+++ net/sched/cls_matchall.c
@@ -160,7 +160,7 @@ static const struct nla_policy mall_policy[TCA_MATCHALL_MAX + 1] = {
 };
 
 static int mall_change(struct net *net, struct sk_buff *in_skb,
-		       struct tcf_proto *tp, unsigned long base,
+		       struct tcf_proto *tp, uintptr_t base,
 		       u32 handle, struct nlattr **tca,
 		       void **arg, u32 flags,
 		       struct netlink_ext_ack *extack)
@@ -378,8 +378,8 @@ nla_put_failure:
 	return -1;
 }
 
-static void mall_bind_class(void *fh, u32 classid, unsigned long cl, void *q,
-			    unsigned long base)
+static void mall_bind_class(void *fh, u32 classid, uintptr_t cl, void *q,
+			    uintptr_t base)
 {
 	struct cls_mall_head *head = fh;
