+++ net/sched/cls_route.c
@@ -474,7 +474,7 @@ static int route4_set_parms(struct net *net, struct tcf_proto *tp,
 }
 
 static int route4_change(struct net *net, struct sk_buff *in_skb,
-			 struct tcf_proto *tp, unsigned long base, u32 handle,
+			 struct tcf_proto *tp, uintptr_t base, u32 handle,
 			 struct nlattr **tca, void **arg, u32 flags,
 			 struct netlink_ext_ack *extack)
 {
@@ -650,8 +650,8 @@ nla_put_failure:
 	return -1;
 }
 
-static void route4_bind_class(void *fh, u32 classid, unsigned long cl, void *q,
-			      unsigned long base)
+static void route4_bind_class(void *fh, u32 classid, uintptr_t cl, void *q,
+			      uintptr_t base)
 {
 	struct route4_filter *f = fh;
