+++ net/sched/cls_u32.c
@@ -867,7 +867,7 @@ static struct tc_u_knode *u32_init_knode(struct net *net, struct tcf_proto *tp,
 }
 
 static int u32_change(struct net *net, struct sk_buff *in_skb,
-		      struct tcf_proto *tp, unsigned long base, u32 handle,
+		      struct tcf_proto *tp, uintptr_t base, u32 handle,
 		      struct nlattr **tca, void **arg, u32 flags,
 		      struct netlink_ext_ack *extack)
 {
@@ -1331,8 +1331,8 @@ static int u32_reoffload(struct tcf_proto *tp, bool add, flow_setup_cb_t *cb,
 	return 0;
 }
 
-static void u32_bind_class(void *fh, u32 classid, unsigned long cl, void *q,
-			   unsigned long base)
+static void u32_bind_class(void *fh, u32 classid, uintptr_t cl, void *q,
+			   uintptr_t base)
 {
 	struct tc_u_knode *n = fh;
