+++ net/sched/sch_cake.c
@@ -3075,27 +3075,27 @@ nla_put_failure:
 	return -1;
 }
 
-static struct Qdisc *cake_leaf(struct Qdisc *sch, unsigned long arg)
+static struct Qdisc *cake_leaf(struct Qdisc *sch, uintptr_t arg)
 {
 	return NULL;
 }
 
-static unsigned long cake_find(struct Qdisc *sch, u32 classid)
+static uintptr_t cake_find(struct Qdisc *sch, u32 classid)
 {
 	return 0;
 }
 
-static unsigned long cake_bind(struct Qdisc *sch, unsigned long parent,
+static uintptr_t cake_bind(struct Qdisc *sch, uintptr_t parent,
 			       u32 classid)
 {
 	return 0;
 }
 
-static void cake_unbind(struct Qdisc *q, unsigned long cl)
+static void cake_unbind(struct Qdisc *q, uintptr_t cl)
 {
 }
 
-static struct tcf_block *cake_tcf_block(struct Qdisc *sch, unsigned long cl,
+static struct tcf_block *cake_tcf_block(struct Qdisc *sch, uintptr_t cl,
 					struct netlink_ext_ack *extack)
 {
 	struct cake_sched_data *q = qdisc_priv(sch);
@@ -3105,14 +3105,14 @@ static struct tcf_block *cake_tcf_block(struct Qdisc *sch, unsigned long cl,
 	return q->block;
 }
 
-static int cake_dump_class(struct Qdisc *sch, unsigned long cl,
+static int cake_dump_class(struct Qdisc *sch, uintptr_t cl,
 			   struct sk_buff *skb, struct tcmsg *tcm)
 {
 	tcm->tcm_handle |= TC_H_MIN(cl);
 	return 0;
 }
 
-static int cake_dump_class_stats(struct Qdisc *sch, unsigned long cl,
+static int cake_dump_class_stats(struct Qdisc *sch, uintptr_t cl,
 				 struct gnet_dump *d)
 {
 	struct cake_sched_data *q = qdisc_priv(sch);
@@ -3314,7 +3314,7 @@ static int cake_mq_change(struct Qdisc *sch, struct nlattr *opt,
 	return 0;
 }
 
-static int cake_mq_graft(struct Qdisc *sch, unsigned long cl, struct Qdisc *new,
+static int cake_mq_graft(struct Qdisc *sch, uintptr_t cl, struct Qdisc *new,
 			 struct Qdisc **old, struct netlink_ext_ack *extack)
 {
 	NL_SET_ERR_MSG(extack, "can't replace cake_mq sub-qdiscs");
