+++ net/sched/sch_cbs.c
@@ -475,7 +475,7 @@ nla_put_failure:
 	return -1;
 }
 
-static int cbs_dump_class(struct Qdisc *sch, unsigned long cl,
+static int cbs_dump_class(struct Qdisc *sch, uintptr_t cl,
 			  struct sk_buff *skb, struct tcmsg *tcm)
 {
 	struct cbs_sched_data *q = qdisc_priv(sch);
@@ -489,7 +489,7 @@ static int cbs_dump_class(struct Qdisc *sch, unsigned long cl,
 	return 0;
 }
 
-static int cbs_graft(struct Qdisc *sch, unsigned long arg, struct Qdisc *new,
+static int cbs_graft(struct Qdisc *sch, uintptr_t arg, struct Qdisc *new,
 		     struct Qdisc **old, struct netlink_ext_ack *extack)
 {
 	struct cbs_sched_data *q = qdisc_priv(sch);
@@ -505,14 +505,14 @@ static int cbs_graft(struct Qdisc *sch, unsigned long arg, struct Qdisc *new,
 	return 0;
 }
 
-static struct Qdisc *cbs_leaf(struct Qdisc *sch, unsigned long arg)
+static struct Qdisc *cbs_leaf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct cbs_sched_data *q = qdisc_priv(sch);
 
 	return q->qdisc;
 }
 
-static unsigned long cbs_find(struct Qdisc *sch, u32 classid)
+static uintptr_t cbs_find(struct Qdisc *sch, u32 classid)
 {
 	return 1;
 }
