+++ net/sched/sch_htb.c
@@ -194,7 +194,7 @@ static inline struct htb_class *htb_find(u32 handle, struct Qdisc *sch)
 	return container_of(clc, struct htb_class, common);
 }
 
-static unsigned long htb_search(struct Qdisc *sch, u32 handle)
+static uintptr_t htb_search(struct Qdisc *sch, u32 handle)
 {
 	return (uintptr_t)htb_find(handle, sch);
 }
@@ -1239,7 +1239,7 @@ nla_put_failure:
 	return -1;
 }
 
-static int htb_dump_class(struct Qdisc *sch, unsigned long arg,
+static int htb_dump_class(struct Qdisc *sch, uintptr_t arg,
 			  struct sk_buff *skb, struct tcmsg *tcm)
 {
 	struct htb_class *cl = (struct htb_class *)arg;
@@ -1319,7 +1319,7 @@ static void htb_offload_aggregate_stats(struct htb_sched *q,
 }
 
 static int
-htb_dump_class_stats(struct Qdisc *sch, unsigned long arg, struct gnet_dump *d)
+htb_dump_class_stats(struct Qdisc *sch, uintptr_t arg, struct gnet_dump *d)
 {
 	struct htb_class *cl = (struct htb_class *)arg;
 	struct htb_sched *q = qdisc_priv(sch);
@@ -1440,7 +1440,7 @@ static void htb_offload_move_qdisc(struct Qdisc *sch, struct htb_class *cl_old,
 	}
 }
 
-static int htb_graft(struct Qdisc *sch, unsigned long arg, struct Qdisc *new,
+static int htb_graft(struct Qdisc *sch, uintptr_t arg, struct Qdisc *new,
 		     struct Qdisc **old, struct netlink_ext_ack *extack)
 {
 	struct netdev_queue *dev_queue = sch->dev_queue;
@@ -1477,13 +1477,13 @@ static int htb_graft(struct Qdisc *sch, unsigned long arg, struct Qdisc *new,
 	return 0;
 }
 
-static struct Qdisc *htb_leaf(struct Qdisc *sch, unsigned long arg)
+static struct Qdisc *htb_leaf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct htb_class *cl = (struct htb_class *)arg;
 	return !cl->level ? cl->leaf.q : NULL;
 }
 
-static void htb_qlen_notify(struct Qdisc *sch, unsigned long arg)
+static void htb_qlen_notify(struct Qdisc *sch, uintptr_t arg)
 {
 	struct htb_class *cl = (struct htb_class *)arg;
 
@@ -1689,7 +1689,7 @@ static void htb_destroy(struct Qdisc *sch)
 	kfree(q->direct_qdiscs);
 }
 
-static int htb_delete(struct Qdisc *sch, unsigned long arg,
+static int htb_delete(struct Qdisc *sch, uintptr_t arg,
 		      struct netlink_ext_ack *extack)
 {
 	struct htb_sched *q = qdisc_priv(sch);
@@ -1757,7 +1757,7 @@ static int htb_delete(struct Qdisc *sch, unsigned long arg,
 
 static int htb_change_class(struct Qdisc *sch, u32 classid,
 			    u32 parentid, struct nlattr **tca,
-			    unsigned long *arg, struct netlink_ext_ack *extack)
+			    uintptr_t *arg, struct netlink_ext_ack *extack)
 {
 	int err = -EINVAL;
 	struct htb_sched *q = qdisc_priv(sch);
@@ -2068,7 +2068,7 @@ failure:
 	return err;
 }
 
-static struct tcf_block *htb_tcf_block(struct Qdisc *sch, unsigned long arg,
+static struct tcf_block *htb_tcf_block(struct Qdisc *sch, uintptr_t arg,
 				       struct netlink_ext_ack *extack)
 {
 	struct htb_sched *q = qdisc_priv(sch);
@@ -2077,7 +2077,7 @@ static struct tcf_block *htb_tcf_block(struct Qdisc *sch, unsigned long arg,
 	return cl ? cl->block : q->block;
 }
 
-static unsigned long htb_bind_filter(struct Qdisc *sch, unsigned long parent,
+static uintptr_t htb_bind_filter(struct Qdisc *sch, uintptr_t parent,
 				     u32 classid)
 {
 	struct htb_class *cl = htb_find(classid, sch);
@@ -2096,7 +2096,7 @@ static unsigned long htb_bind_filter(struct Qdisc *sch, unsigned long parent,
 	return (uintptr_t)cl;
 }
 
-static void htb_unbind_filter(struct Qdisc *sch, unsigned long arg)
+static void htb_unbind_filter(struct Qdisc *sch, uintptr_t arg)
 {
 	struct htb_class *cl = (struct htb_class *)arg;
